Добавляет разбивку по модулям
